refactor: move ue_math.h to game/shared
